+++ tt_metal/hw/ckernels/blackhole/metal/llk_api/experimental/llk_sfpu/llk_math_eltwise_unary_sfpu_topk_xl.h
@@ -6,7 +6,7 @@
 
 #include "llk_math_eltwise_unary_sfpu_init.h"
 #include "llk_math_eltwise_unary_sfpu_params.h"
-#include "sfpu/ckernel_sfpu_topk_xl.h"
+#include "sfpu/experimental/ckernel_sfpu_topk_xl.h"
 
 namespace ckernel {
